Submission Files/requirements.txt
imageio==2.34.1
matplotlib==3.8.4
numpy==1.26.4
Pillow==10.3.0
torch==2.3.0
torchvision==0.18.0
tqdm==4.66.2
opencv-python
scikit-image